+++ src/standard/necessary_tiles.rs
@@ -166,7 +166,7 @@ pub(in super::super) fn calculate_necessary_tiles(bingpai: &Bingpai) -> (u8, Til
     update_dp_final(&mut entry0, &entry3);
 
     let n = 5 + bingpai.num_required_bingpai_mianzi() as usize;
-    (entry0.numbers[n], entry0.tiles[n])
+    (entry0.numbers[n] as u8, entry0.tiles[n])
 }
 
 pub(in super::super) fn calculate_necessary_tiles_3p(bingpai: &Bingpai3p) -> (u8, TileFlags) {
@@ -217,7 +217,7 @@ pub(in super::super) fn calculate_necessary_tiles_3p(bingpai: &Bingpai3p) -> (u8
     update_dp_final(&mut entry0, &entry3);
 
     let n = 5 + bingpai.num_required_bingpai_mianzi() as usize;
-    (entry0.numbers[n], entry0.tiles[n])
+    (entry0.numbers[n] as u8, entry0.tiles[n])
 }
 
 #[cfg(test)]
